+++ code/src/sixsq/nuvla/server/resources/user_email_invitation.clj
@@ -34,7 +34,7 @@ using an email address.
 
 (defmethod p/tpl->user email-invitation/registration-method
   [resource request]
-  (password-utils/create-user-map resource))
+  [nil (password-utils/create-user-map resource)])
 
 
 ;;